Refactor: Migrate currency from float to integer pennies
Global migration of monetary values from `float` to `int` (pennies) to ensure zero-sum integrity and eliminate floating-point errors.

Key changes:
- Created `modules/finance/utils/currency_math.py` with `round_to_pennies` (Banker's Rounding).
- Updated core DTOs (`MoneyDTO`, `LoanStateDTO`, etc.) to use `int` and `_pennies` suffix.
- Updated interfaces (`IFinancialAgent`, `IBank`, `ISettlementSystem`) to enforce `int`.
- Updated `Wallet` to store `Dict[CurrencyCode, int]`.
- Hardened `SettlementSystem` to strictly enforce `int` transfers and zero-leak logic.
- Updated Core Agents (`Household`, `Firm`, `Government`, `Bank`) to use integer state.
- Updated Engines (`DebtServicingEngine`, `BudgetEngine`, `FinanceEngine`, `FiscalEngine`, `TaxService`) to perform calculation with `Decimal` and output `int`.
- Updated `SalesEngine` to use `marketing_budget_pennies`.
- Updated unit tests (`test_finance_engines.py`, `test_bank.py`, `test_system.py`) to reflect changes.

+++ modules/household/api.py
@@ -27,8 +27,8 @@ class PrioritizedNeed:
 @dataclass
 class BudgetPlan:
     """A concrete allocation of funds for the current tick."""
-    allocations: Dict[str, float]  # e.g., {"food": 100, "housing": 500, "savings": 50}
-    discretionary_spending: float
+    allocations: Dict[str, int]  # MIGRATION: int pennies. e.g., {"food": 10000}
+    discretionary_spending: int # MIGRATION: int pennies.
     orders: List[Order] = field(default_factory=list) # Orders approved by budget
 
 @dataclass
@@ -39,8 +39,8 @@ class HousingActionDTO:
     """
     action_type: str  # "INITIATE_PURCHASE", "MAKE_RENTAL_OFFER", "STAY"
     property_id: Optional[str] = None
-    offer_price: float = 0.0
-    down_payment_amount: float = 0.0
+    offer_price: int = 0
+    down_payment_amount: int = 0
     buyer_id: Optional[int] = None
 
 # --- Engine Input DTOs ---
